Updating the README with project information
The motivation behind the CSCE 625 Self Evaluation project is to help students in assessing their
preparation for the graduate-level Artificial Intelligence course. This project can assist students in
understanding the course prerequisites. By using this evaluation portal, the students can make an
informed decision as to whether they are sufficiently prepared to register for the class.

The requirements for the project were provided by Dr Duncan Walker, Professor, Texas A&M
University. The key stakeholders of this project are prospective students, teaching assistants, and
the course instructors.